rename views/RootView to views/rootview
