+++ src/actions/profile.js
@@ -1,7 +1,7 @@
 import fetch from 'isomorphic-fetch';
 import { browserHistory } from 'react-router'
 
-import { store } from '../../views/RootView'
+import { store } from '../../views/rootview'
 
 export const REQUEST_USERS = 'REQUEST_USERS';
 export const RECEIVE_USERS = 'RECEIVE_USERS';
